src/main/java/de/tudarmstadt/aiphes/mdswriter/data/BestNuggetManager.java
/*******************************************************************************
 * Copyright 2016
 * Ubiquitous Knowledge Processing (UKP) Lab
 * Technische Universität Darmstadt
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *   http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 ******************************************************************************/
package de.tudarmstadt.aiphes.mdswriter.data;

import java.sql.PreparedStatement;
import java.sql.SQLException;
import java.util.ArrayList;
import java.util.HashSet;
import java.util.List;
import java.util.Set;

import javax.websocket.Session;

import de.tudarmstadt.aiphes.mdswriter.auth.AuthManager;
import de.tudarmstadt.aiphes.mdswriter.auth.User;
import de.tudarmstadt.aiphes.mdswriter.db.DBConnection;
import de.tudarmstadt.aiphes.mdswriter.db.DBManager;
import de.tudarmstadt.aiphes.mdswriter.doc.DocumentManager;
import de.tudarmstadt.aiphes.mdswriter.doc.DocumentSet;

public class BestNuggetManager extends NuggetManager<NuggetClusterItem> {

	public BestNuggetManager(User user, DocumentSet docSet) throws SQLException {
		super(user, docSet);

		// Load the best nuggets ordered by clusterPos.
		DBConnection connection = DBManager.getConnection();
		try {
			nuggets = Nugget.listBest(connection, user, docSet);
			createItemList();
		} finally {
			connection.close();
		}
	}

	protected BestNuggetManager(final List<Nugget> nuggets) {
		super(nuggets);
	}

	protected void createItemList() {
		items = new ArrayList<NuggetClusterItem>();
		int prevGroup = 0;
		int newGroup = 0;
		for (Nugget nugget : nuggets) {
			if (nugget.getClusterId() != prevGroup) {
				newGroup++;
				items.add(createGroupItem(newGroup, nugget.getClusterName()));
				prevGroup = nugget.getClusterId();
			}
			items.add(createNuggetItem(nugget, newGroup, nugget.getClusterName()));
		}
	}

	protected NuggetClusterItem createGroupItem(final int groupId,
			final String groupTitle) {
		return new NuggetClusterItem(groupId, groupTitle);
	}

	protected NuggetClusterItem createNuggetItem(final Nugget nugget,
			final int groupId, final String groupTitle) {
		return new NuggetClusterItem(nugget, groupId, groupTitle);
	}

	protected void doSaveItems(final DBConnection connection,
			final Set<NuggetClusterItem> itemsToSave) throws SQLException {
		if (itemsToSave == null || itemsToSave.isEmpty())
			return;

		if (DEBUG_MODE)
			System.out.println("SAVING TO DB!");
		PreparedStatement stmt = connection.prepareStatement(
				"UPDATE " + Nugget.table().getName()
				+ " SET cluster_pos = ?, cluster_id = ?, cluster_name = ?"
				+ " WHERE id = ?");
		try {
			for (NuggetClusterItem item : itemsToSave)
				item.save(stmt);
		} finally {
			stmt.close();
		}
	}

	public void saveCoreference(final int nuggetId, final String text)
			throws SQLException {
		Nugget nugget = findNugget(nuggetId);
		if (nugget == null)
			throw new NullPointerException("Invalid nugget id: " + nuggetId);

		nugget.setTextCoreference(text);

		DBConnection connection = DBManager.getConnection();
		try {
			PreparedStatement stmt = connection.prepareStatement(
					"UPDATE " + Nugget.table().getName()
					+ " SET nugget_coref = ? WHERE id = ?");
			try {
				stmt.setString(1, text);
				stmt.setInt(2, nuggetId);
				stmt.execute();
			} finally {
				stmt.close();
			}
		} finally {
			connection.close();
		}
	}

	public void saveRevision(final int nuggetId, final String text)
			throws SQLException {
		Nugget nugget = findNugget(nuggetId);
		if (nugget == null)
			throw new NullPointerException("Invalid nugget id: " + nuggetId);

		nugget.setTextRevised(text);

		DBConnection connection = DBManager.getConnection();
		try {
			PreparedStatement stmt = connection.prepareStatement(
					"UPDATE " + Nugget.table().getName()
					+ " SET nugget_revised = ? WHERE id = ?");
			try {
				stmt.setString(1, text);
				stmt.setInt(2, nuggetId);
				stmt.execute();
			} finally {
				stmt.close();
			}
		} finally {
			connection.close();
		}
	}

	public void renameGroup(final int index, final String title)
			 throws SQLException {
		saveItems(doRenameGroup(index, title));
		if (DEBUG_MODE)
			dumpItems();
	}

	protected Set<NuggetClusterItem> doRenameGroup(final int index,
			final String title) {
		// Ensure the specified item is a group header.
		if (!items.get(index).isGroupHeader())
			throw new IllegalArgumentException("Item at index " + index + " is not a group header");

		// Correct the group ids of all items after the new header.
		Set<NuggetClusterItem> itemsToSave = new HashSet<NuggetClusterItem>();
		for (int i = index; i < items.size(); i++) {
			NuggetClusterItem item = items.get(i);
			if (i > index && item.isGroupHeader())
				break;

			if (item.setGroupTitle(title))
				itemsToSave.add(item);
		}
		return itemsToSave;
	}

	protected void dumpItems() {
		System.out.println();
		int j = 0;
		for (NuggetClusterItem i : items) {
			System.out.printf("%4d  ", j);
			if (i.isGroupHeader())
				System.out.printf("[---] Group %3d %s\n",
						i.getGroupId(), i.getGroupTitle());
			else
				System.out.printf("[%3d] Cluster %3d > %3d  /  ClusterPos %3d > %3d  /  \"%s\" > \"%s\"\n",
						i.getNugget().getId(),
						i.getGroupId(), i.getNugget().getClusterId(),
						i.getNugget().getClusterPos(), i.computeNewPos(j),
						i.getGroupTitle(),
						i.getNugget().getClusterName());
			j++;
		}
		System.out.println();
	}

	protected void dumpNuggets() {
		for (Nugget n : nuggets)
			System.out.println(n.getId() + " " + n.getClusterId() + " " + n.getClusterPos());
		System.out.println();
	}


	public static BestNuggetManager get(final Session session)
			throws SQLException {
		BestNuggetManager result = (BestNuggetManager) session.getUserProperties().get("bestNuggets");
		if (result == null) {
			DocumentSet docSet = DocumentManager.getSessionDocSet(session);
			if (docSet == null)
				throw new NullPointerException("Session docSet missing");

			User user = AuthManager.getAccessToken(session).getUser();
			result = new BestNuggetManager(user, docSet);
			session.getUserProperties().put("bestNuggets", result);
		}
		return result;
	}

}
