src/test/java/de/tudarmstadt/aiphes/mdswriter/data/BestNuggetManagerTest.java
/*******************************************************************************
 * Copyright 2016
 * Ubiquitous Knowledge Processing (UKP) Lab
 * Technische Universität Darmstadt
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *   http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 ******************************************************************************/
package de.tudarmstadt.aiphes.mdswriter.data;

import java.util.ArrayList;
import java.util.Arrays;
import java.util.Iterator;
import java.util.List;
import java.util.Set;
import java.util.TreeSet;

import de.tudarmstadt.aiphes.mdswriter.data.BestNuggetManager;
import de.tudarmstadt.aiphes.mdswriter.data.Nugget;
import de.tudarmstadt.aiphes.mdswriter.data.NuggetClusterItem;
import junit.framework.TestCase;

public class BestNuggetManagerTest extends TestCase {

	protected static class TestNugget extends Nugget {

		public TestNugget(final int id, final int clusterId,
				final String clusterName, final int clusterPos) {
			super(id, null, null, null, "", null, -1, -1, 0, 0, false,
					"", "", clusterPos, clusterId, clusterName, null);
		}

	}

	public void testMakeConsistentClusters() {
		List<Nugget> nuggets = new ArrayList<Nugget>();
		nuggets.add(new TestNugget(101, 0, "x", 0));
		nuggets.add(new TestNugget(102, 0, null, 0));
		nuggets.add(new TestNugget(103, 0, "A", 0));
		nuggets.add(new TestNugget(104, 0, "A", 0));
		nuggets.add(new TestNugget(105, 0, "x", 0));

		BestNuggetManager nm = new BestNuggetManager(nuggets);
		Set<NuggetClusterItem> itemsToSave = nm.doMakeConsistent();
		Iterator<Nugget> iter = nm.getNuggets().iterator();
		assertNugget(101, 0, null, 1, iter.next());
		assertNugget(102, 0, null, 2, iter.next());
		assertNugget(103, 0, null, 3, iter.next());
		assertNugget(104, 0, null, 4, iter.next());
		assertNugget(105, 0, null, 5, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 101, 102, 103, 104, 105);

		itemsToSave = nm.doMakeConsistent();
		iter = nm.getNuggets().iterator();
		assertNugget(101, 0, null, 1, iter.next());
		assertNugget(102, 0, null, 2, iter.next());
		assertNugget(103, 0, null, 3, iter.next());
		assertNugget(104, 0, null, 4, iter.next());
		assertNugget(105, 0, null, 5, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave);

		nuggets = new ArrayList<Nugget>();
		nuggets.add(new TestNugget(101, 2, "B", 2));
		nuggets.add(new TestNugget(102, 1, "A", 3));
		nuggets.add(new TestNugget(103, 1, "B", 5));
		nuggets.add(new TestNugget(104, 1, "C", 5));
		nuggets.add(new TestNugget(105, 3, "C", 5));
		nuggets.add(new TestNugget(106, 1, "C", 9));

		nm = new BestNuggetManager(nuggets);
		itemsToSave = nm.doMakeConsistent();
		iter = nm.getNuggets().iterator();
		assertNugget(101, 1, "B", 1, iter.next());
		assertNugget(102, 2, "A", 2, iter.next());
		assertNugget(103, 2, "A", 3, iter.next());
		assertNugget(104, 2, "A", 4, iter.next());
		assertNugget(105, 3, "C", 5, iter.next());
		assertNugget(106, 4, "C", 6, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 101, 102, 103, 104, 106);

		itemsToSave = nm.doMakeConsistent();
		iter = nm.getNuggets().iterator();
		assertNugget(101, 1, "B", 1, iter.next());
		assertNugget(102, 2, "A", 2, iter.next());
		assertNugget(103, 2, "A", 3, iter.next());
		assertNugget(104, 2, "A", 4, iter.next());
		assertNugget(105, 3, "C", 5, iter.next());
		assertNugget(106, 4, "C", 6, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave);
	}

	public void testAddGroup() {
		List<Nugget> nuggets = new ArrayList<Nugget>();
		nuggets.add(new TestNugget(101, 0, null, 1));
		nuggets.add(new TestNugget(102, 0, null, 2));
		nuggets.add(new TestNugget(103, 0, null, 3));
		nuggets.add(new TestNugget(104, 0, null, 4));
		nuggets.add(new TestNugget(105, 0, null, 5));

		BestNuggetManager nm = new BestNuggetManager(nuggets);
		Set<NuggetClusterItem> itemsToSave = nm.doAddGroup(3, "A");
		Iterator<Nugget> iter = nm.getNuggets().iterator();
		assertNugget(101, 0, null, 1, iter.next());
		assertNugget(102, 0, null, 2, iter.next());
		assertNugget(103, 0, null, 3, iter.next());
		assertNugget(104, 1, "A", 4, iter.next());
		assertNugget(105, 1, "A", 5, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 104, 105);

		itemsToSave = nm.doAddGroup(1, "B");
		iter = nm.getNuggets().iterator();
		assertNugget(101, 0, null, 1, iter.next());
		assertNugget(102, 1, "B", 2, iter.next());
		assertNugget(103, 1, "B", 3, iter.next());
		assertNugget(104, 2, "A", 4, iter.next());
		assertNugget(105, 2, "A", 5, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 102, 103, 104, 105);

		itemsToSave = nm.doAddGroup(0, "C");
		iter = nm.getNuggets().iterator();
		assertNugget(101, 1, "C", 1, iter.next());
		assertNugget(102, 2, "B", 2, iter.next());
		assertNugget(103, 2, "B", 3, iter.next());
		assertNugget(104, 3, "A", 4, iter.next());
		assertNugget(105, 3, "A", 5, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 101, 102, 103, 104, 105);

		itemsToSave = nm.doAddGroup(7, "D");
		iter = nm.getNuggets().iterator();
		assertNugget(101, 1, "C", 1, iter.next());
		assertNugget(102, 2, "B", 2, iter.next());
		assertNugget(103, 2, "B", 3, iter.next());
		assertNugget(104, 3, "A", 4, iter.next());
		assertNugget(105, 4, "D", 5, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 105);

		itemsToSave = nm.doAddGroup(9, "E");
		iter = nm.getNuggets().iterator();
		assertNugget(101, 1, "C", 1, iter.next());
		assertNugget(102, 2, "B", 2, iter.next());
		assertNugget(103, 2, "B", 3, iter.next());
		assertNugget(104, 3, "A", 4, iter.next());
		assertNugget(105, 4, "D", 5, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave);

		try {
			nm.doAddGroup(11, "F");
			fail("IndexOutOfBoundsException expected");
		} catch (IndexOutOfBoundsException e) {}
	}

	public void testRemoveGroup() {
		List<Nugget> nuggets = new ArrayList<Nugget>();
		nuggets.add(new TestNugget(101, 1, "A", 1));
		nuggets.add(new TestNugget(102, 2, "B", 2));
		nuggets.add(new TestNugget(103, 2, "B", 3));
		nuggets.add(new TestNugget(104, 3, "C", 4));
		nuggets.add(new TestNugget(105, 4, "D", 5));

		BestNuggetManager nm = new BestNuggetManager(nuggets);
		try {
			nm.doRemoveGroup(9);
			fail("IndexOutOfBoundsException expected");
		} catch (IndexOutOfBoundsException e) {}

		try {
			nm.doRemoveGroup(8);
			fail("IllegalArgumentException expected");
		} catch (IllegalArgumentException e) {}

		Set<NuggetClusterItem> itemsToSave = nm.doRemoveGroup(5);
		Iterator<Nugget> iter = nm.getNuggets().iterator();
		assertNugget(101, 1, "A", 1, iter.next());
		assertNugget(102, 2, "B", 2, iter.next());
		assertNugget(103, 2, "B", 3, iter.next());
		assertNugget(104, 2, "B", 4, iter.next());
		assertNugget(105, 3, "D", 5, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 104, 105);

		itemsToSave = nm.doRemoveGroup(0);
		iter = nm.getNuggets().iterator();
		assertNugget(101, 0, null, 1, iter.next());
		assertNugget(102, 1, "B", 2, iter.next());
		assertNugget(103, 1, "B", 3, iter.next());
		assertNugget(104, 1, "B", 4, iter.next());
		assertNugget(105, 2, "D", 5, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 101, 102, 103, 104, 105);

		itemsToSave = nm.doRemoveGroup(1);
		iter = nm.getNuggets().iterator();
		assertNugget(101, 0, null, 1, iter.next());
		assertNugget(102, 0, null, 2, iter.next());
		assertNugget(103, 0, null, 3, iter.next());
		assertNugget(104, 0, null, 4, iter.next());
		assertNugget(105, 1, "D", 5, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 102, 103, 104, 105);

		itemsToSave = nm.doRemoveGroup(4);
		iter = nm.getNuggets().iterator();
		assertNugget(101, 0, null, 1, iter.next());
		assertNugget(102, 0, null, 2, iter.next());
		assertNugget(103, 0, null, 3, iter.next());
		assertNugget(104, 0, null, 4, iter.next());
		assertNugget(105, 0, null, 5, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 105);
	}

	public void testMoveGroup() {
		List<Nugget> nuggets = new ArrayList<Nugget>();
		nuggets.add(new TestNugget(101, 1, "A", 1));
		nuggets.add(new TestNugget(102, 2, "B", 2));
		nuggets.add(new TestNugget(103, 2, "B", 3));
		nuggets.add(new TestNugget(104, 2, "B", 4));
		nuggets.add(new TestNugget(105, 3, "C", 5));
		nuggets.add(new TestNugget(106, 3, "C", 6));
		nuggets.add(new TestNugget(107, 4, "D", 7));
		nuggets.add(new TestNugget(108, 4, "D", 8));

		BestNuggetManager nm = new BestNuggetManager(nuggets);
		Set<NuggetClusterItem> itemsToSave = nm.doMoveGroup(6, 0);
		Iterator<Nugget> iter = nm.getNuggets().iterator();
		assertNugget(101, 2, "A", 1, iter.next());
		assertNugget(102, 3, "B", 2, iter.next());
		assertNugget(103, 3, "B", 3, iter.next());
		assertNugget(104, 3, "B", 4, iter.next());
		assertNugget(105, 3, "B", 5, iter.next());
		assertNugget(106, 3, "B", 6, iter.next());
		assertNugget(107, 4, "D", 7, iter.next());
		assertNugget(108, 4, "D", 8, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 101, 102, 103, 104, 105, 106);

		itemsToSave = nm.doMoveGroup(1, 4);
		iter = nm.getNuggets().iterator();
		assertNugget(101, 1, "C", 1, iter.next());
		assertNugget(102, 3, "A", 2, iter.next());
		assertNugget(103, 3, "A", 3, iter.next());
		assertNugget(104, 3, "A", 4, iter.next());
		assertNugget(105, 3, "A", 5, iter.next());
		assertNugget(106, 3, "A", 6, iter.next());
		assertNugget(107, 4, "D", 7, iter.next());
		assertNugget(108, 4, "D", 8, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 101, 102, 103, 104, 105, 106);

		itemsToSave = nm.doMoveGroup(3, 5);
		iter = nm.getNuggets().iterator();
		assertNugget(101, 1, "C", 1, iter.next());
		assertNugget(102, 2, "B", 2, iter.next());
		assertNugget(103, 3, "A", 3, iter.next());
		assertNugget(104, 3, "A", 4, iter.next());
		assertNugget(105, 3, "A", 5, iter.next());
		assertNugget(106, 3, "A", 6, iter.next());
		assertNugget(107, 4, "D", 7, iter.next());
		assertNugget(108, 4, "D", 8, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 102);

		itemsToSave = nm.doMoveGroup(0, 12);
		iter = nm.getNuggets().iterator();
		assertNugget(101, 0, null, 1, iter.next());
		assertNugget(102, 1, "B", 2, iter.next());
		assertNugget(103, 2, "A", 3, iter.next());
		assertNugget(104, 2, "A", 4, iter.next());
		assertNugget(105, 2, "A", 5, iter.next());
		assertNugget(106, 2, "A", 6, iter.next());
		assertNugget(107, 3, "D", 7, iter.next());
		assertNugget(108, 3, "D", 8, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 101, 102, 103, 104, 105, 106, 107, 108);

		itemsToSave = nm.doMoveGroup(8, 6);
		iter = nm.getNuggets().iterator();
		assertNugget(101, 0, null, 1, iter.next());
		assertNugget(102, 1, "B", 2, iter.next());
		assertNugget(103, 2, "A", 3, iter.next());
		assertNugget(104, 2, "A", 4, iter.next());
		assertNugget(105, 3, "D", 5, iter.next());
		assertNugget(106, 3, "D", 6, iter.next());
		assertNugget(107, 3, "D", 7, iter.next());
		assertNugget(108, 3, "D", 8, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 105, 106);
	}

	public void testMoveNugget() {
		List<Nugget> nuggets = new ArrayList<Nugget>();
		nuggets.add(new TestNugget(101, 1, "A", 1));
		nuggets.add(new TestNugget(102, 2, "B", 2));
		nuggets.add(new TestNugget(103, 2, "B", 3));
		nuggets.add(new TestNugget(104, 2, "B", 4));
		nuggets.add(new TestNugget(105, 3, "C", 5));
		nuggets.add(new TestNugget(106, 3, "C", 6));
		nuggets.add(new TestNugget(107, 4, "D", 7));
		nuggets.add(new TestNugget(108, 4, "D", 8));

		BestNuggetManager nm = new BestNuggetManager(nuggets);
		Set<NuggetClusterItem> itemsToSave = nm.doMoveNugget(3, 6);
		Iterator<Nugget> iter = nm.getNuggets().iterator();
		assertNugget(101, 1, "A", 1, iter.next());
		assertNugget(103, 2, "B", 2, iter.next());
		assertNugget(104, 2, "B", 3, iter.next());
		assertNugget(102, 2, "B", 4, iter.next());
		assertNugget(105, 3, "C", 5, iter.next());
		assertNugget(106, 3, "C", 6, iter.next());
		assertNugget(107, 4, "D", 7, iter.next());
		assertNugget(108, 4, "D", 8, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 103, 104, 102);

		itemsToSave = nm.doMoveNugget(5, 3);
		iter = nm.getNuggets().iterator();
		assertNugget(101, 1, "A", 1, iter.next());
		assertNugget(102, 2, "B", 2, iter.next());
		assertNugget(103, 2, "B", 3, iter.next());
		assertNugget(104, 2, "B", 4, iter.next());
		assertNugget(105, 3, "C", 5, iter.next());
		assertNugget(106, 3, "C", 6, iter.next());
		assertNugget(107, 4, "D", 7, iter.next());
		assertNugget(108, 4, "D", 8, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 102, 103, 104);

		itemsToSave = nm.doMoveNugget(3, 2);
		iter = nm.getNuggets().iterator();
		assertNugget(101, 1, "A", 1, iter.next());
		assertNugget(102, 1, "A", 2, iter.next());
		assertNugget(103, 2, "B", 3, iter.next());
		assertNugget(104, 2, "B", 4, iter.next());
		assertNugget(105, 3, "C", 5, iter.next());
		assertNugget(106, 3, "C", 6, iter.next());
		assertNugget(107, 4, "D", 7, iter.next());
		assertNugget(108, 4, "D", 8, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 102);

		itemsToSave = nm.doMoveNugget(4, 2);
		iter = nm.getNuggets().iterator();
		assertNugget(101, 1, "A", 1, iter.next());
		assertNugget(103, 1, "A", 2, iter.next());
		assertNugget(102, 1, "A", 3, iter.next());
		assertNugget(104, 2, "B", 4, iter.next());
		assertNugget(105, 3, "C", 5, iter.next());
		assertNugget(106, 3, "C", 6, iter.next());
		assertNugget(107, 4, "D", 7, iter.next());
		assertNugget(108, 4, "D", 8, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 103, 102);

		itemsToSave = nm.doMoveNugget(5, 10);
		iter = nm.getNuggets().iterator();
		assertNugget(101, 1, "A", 1, iter.next());
		assertNugget(103, 1, "A", 2, iter.next());
		assertNugget(102, 1, "A", 3, iter.next());
		assertNugget(105, 3, "C", 4, iter.next());
		assertNugget(106, 3, "C", 5, iter.next());
		assertNugget(104, 4, "D", 6, iter.next());
		assertNugget(107, 4, "D", 7, iter.next());
		assertNugget(108, 4, "D", 8, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 105, 106, 104);

		itemsToSave = nm.doMoveNugget(6, 0);
		iter = nm.getNuggets().iterator();
		assertNugget(105, 0, null, 1, iter.next());
		assertNugget(101, 1, "A", 2, iter.next());
		assertNugget(103, 1, "A", 3, iter.next());
		assertNugget(102, 1, "A", 4, iter.next());
		assertNugget(106, 3, "C", 5, iter.next());
		assertNugget(104, 4, "D", 6, iter.next());
		assertNugget(107, 4, "D", 7, iter.next());
		assertNugget(108, 4, "D", 8, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 105, 101, 102, 103);

		itemsToSave = nm.doMoveNugget(7, 12);
		iter = nm.getNuggets().iterator();
		assertNugget(105, 0, null, 1, iter.next());
		assertNugget(101, 1, "A", 2, iter.next());
		assertNugget(103, 1, "A", 3, iter.next());
		assertNugget(102, 1, "A", 4, iter.next());
		assertNugget(104, 4, "D", 5, iter.next());
		assertNugget(107, 4, "D", 6, iter.next());
		assertNugget(108, 4, "D", 7, iter.next());
		assertNugget(106, 4, "D", 8, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 104, 107, 108, 106);
	}

	public void testMoveItem() {
		List<Nugget> nuggets = new ArrayList<Nugget>();
		nuggets.add(new TestNugget(101, 0, null, 1));
		nuggets.add(new TestNugget(102, 0, null, 2));
		nuggets.add(new TestNugget(103, 1, "A", 3));
		nuggets.add(new TestNugget(104, 2, "B", 4));
		nuggets.add(new TestNugget(105, 2, "B", 5));

		BestNuggetManager nm = new BestNuggetManager(nuggets);
		try {
			nm.doMoveItem(-1, 0);
			fail("IndexOutOfBoundsException expected");
		} catch (IndexOutOfBoundsException e) {}
		try {
			nm.doMoveItem(7, 0);
			fail("IndexOutOfBoundsException expected");
		} catch (IndexOutOfBoundsException e) {}

		try {
			nm.doMoveItem(0, -1);
			fail("IndexOutOfBoundsException expected");
		} catch (IndexOutOfBoundsException e) {}
		try {
			nm.doMoveItem(0, 8);
			fail("IndexOutOfBoundsException expected");
		} catch (IndexOutOfBoundsException e) {}
	}

	public void testRenameGroup() {
		List<Nugget> nuggets = new ArrayList<Nugget>();
		nuggets.add(new TestNugget(101, 0, null, 1));
		nuggets.add(new TestNugget(102, 1, "A", 2));
		nuggets.add(new TestNugget(103, 1, "A", 3));
		nuggets.add(new TestNugget(104, 1, "A", 4));
		nuggets.add(new TestNugget(105, 2, "B", 5));
		nuggets.add(new TestNugget(106, 2, "B", 6));
		nuggets.add(new TestNugget(107, 3, "C", 7));
		nuggets.add(new TestNugget(108, 4, "D", 8));

		BestNuggetManager nm = new BestNuggetManager(nuggets);
		nm.doMoveItem(9, 11);
		Set<NuggetClusterItem> itemsToSave = nm.doRenameGroup(1, "Ax");
		Iterator<Nugget> iter = nm.getNuggets().iterator();
		assertNugget(101, 0, null, 1, iter.next());
		assertNugget(102, 1, "Ax", 2, iter.next());
		assertNugget(103, 1, "Ax", 3, iter.next());
		assertNugget(104, 1, "Ax", 4, iter.next());
		assertNugget(105, 2, "B", 5, iter.next());
		assertNugget(106, 2, "B", 6, iter.next());
		assertNugget(107, 4, "D", 7, iter.next());
		assertNugget(108, 4, "D", 8, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 102, 103, 104);

		itemsToSave = nm.doRenameGroup(5, "By");
		iter = nm.getNuggets().iterator();
		assertNugget(101, 0, null, 1, iter.next());
		assertNugget(102, 1, "Ax", 2, iter.next());
		assertNugget(103, 1, "Ax", 3, iter.next());
		assertNugget(104, 1, "Ax", 4, iter.next());
		assertNugget(105, 2, "By", 5, iter.next());
		assertNugget(106, 2, "By", 6, iter.next());
		assertNugget(107, 4, "D", 7, iter.next());
		assertNugget(108, 4, "D", 8, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 105, 106);

		itemsToSave = nm.doRenameGroup(8, "Ck");
		iter = nm.getNuggets().iterator();
		assertNugget(101, 0, null, 1, iter.next());
		assertNugget(102, 1, "Ax", 2, iter.next());
		assertNugget(103, 1, "Ax", 3, iter.next());
		assertNugget(104, 1, "Ax", 4, iter.next());
		assertNugget(105, 2, "By", 5, iter.next());
		assertNugget(106, 2, "By", 6, iter.next());
		assertNugget(107, 4, "D", 7, iter.next());
		assertNugget(108, 4, "D", 8, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave);

		itemsToSave = nm.doRenameGroup(9, "Dz");
		iter = nm.getNuggets().iterator();
		assertNugget(101, 0, null, 1, iter.next());
		assertNugget(102, 1, "Ax", 2, iter.next());
		assertNugget(103, 1, "Ax", 3, iter.next());
		assertNugget(104, 1, "Ax", 4, iter.next());
		assertNugget(105, 2, "By", 5, iter.next());
		assertNugget(106, 2, "By", 6, iter.next());
		assertNugget(107, 4, "Dz", 7, iter.next());
		assertNugget(108, 4, "Dz", 8, iter.next());
		assertFalse(iter.hasNext());
		assertChanges(itemsToSave, 107, 108);

		try {
			nm.doRenameGroup(-1, "err");
			fail("IndexOutOfBoundsException expected");
		} catch (IndexOutOfBoundsException e) {}
		try {
			nm.doRenameGroup(12, "err");
			fail("IndexOutOfBoundsException expected");
		} catch (IndexOutOfBoundsException e) {}

		try {
			nm.doRenameGroup(0, "err");
			fail("IllegalArgumentException expected");
		} catch (IllegalArgumentException e) {}
		try {
			nm.doRenameGroup(3, "err");
			fail("IllegalArgumentException expected");
		} catch (IllegalArgumentException e) {}
	}

	protected void assertNugget(final int id, final int clusterId,
			final String clusterName, final int clusterPos,
			final Nugget actual) {
		assertEquals(id, actual.getId());
		assertEquals(clusterId, actual.getClusterId());
		assertEquals(clusterName, actual.getClusterName());
		assertEquals(clusterPos, actual.getClusterPos());
	}

	protected void assertChanges(final Set<NuggetClusterItem> itemsToSave,
			final int... expectedIDs) {
		TreeSet<Integer> expected = new TreeSet<Integer>();
		for (int expectedID : expectedIDs)
			expected.add(expectedID);

		TreeSet<Integer> actual = new TreeSet<Integer>();
		for (NuggetClusterItem item : itemsToSave)
			if (!item.isGroupHeader())
				actual.add(item.getNugget().getId());

		assertEquals(Arrays.toString(expected.toArray()),
				Arrays.toString(actual.toArray()));
	}

}
